python/diarization/pyproject.toml
[project]
name = "diarization"
version = "0.1.0"
description = "Speaker diarization using pyannote.audio"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Now using v4.x with locally built torchcodec for aarch64 Linux!
    "pyannote-audio>=4.0.0",
    "typer>=0.9.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "torchvision>=0.15.0",
    "pybind11>=3.0.1",
    "torchcodec",
]

[project.scripts]
diarize = "diarization.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "syrupy>=4.0.0",
]

[tool.uv]

[tool.uv.sources]
torchcodec = { path = "../../.devcontainer/torchcodec-0.9.0a0-cp312-cp312-linux_aarch64.whl" }

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
